moved admin home view styles to same folder as components
