+++ components/Home/AdminHomeView.module.css
@@ -14,7 +14,7 @@
   box-shadow: 0px 2px 14px rgba(0, 0, 0, 0.18);
   border: none;
   cursor: pointer;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
 }
 
 .addIcon {
