+++ components/Home/CreateEventsWizard.module.css
@@ -73,7 +73,7 @@
 
 .label {
   margin: 0px;
-  composes: headlineSmall from "../globals.css";
+  composes: headlineSmall from "../../styles/globals.css";
 }
 
 .classInput {
@@ -93,7 +93,6 @@
   display: flex;
   justify-content: space-between;
   align-items: center;
-  /*width: 375px;*/
   margin-top: 32px;
 }
 
@@ -105,7 +104,7 @@
 
 .checkLabel {
   margin: 0px;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
 }
 
 .row {
@@ -120,7 +119,7 @@
   border-radius: 8px;
   background: var(--gray-4);
   color: var(--black);
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
 }
 
 .dropDownWrapper {
@@ -152,7 +151,7 @@
 
 .dateInput {
   composes: inputField;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
   width: 170px;
   height: 40px;
   margin: 10px 0px;
@@ -162,7 +161,7 @@
 
 .timeInput {
   composes: inputField;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
   width: 105px;
   height: 40px;
   margin: 10px 0px;
@@ -196,9 +195,8 @@
   cursor: pointer;
 
   white-space: nowrap;
-  overflow: hidden;
-  text-overflow: ellipsis;
-  composes: labelLarge from "../globals.css";
+  overflow: auto;
+  composes: labelLarge from "../../styles/globals.css";
 }
 
 .recurrenceInput::placeholder {
@@ -207,7 +205,7 @@
 
 .textInput {
   composes: inputField;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
   width: 648px;
   height: 40px;
   margin: 10px 0px;
