+++ components/Home/RepeatModal.module.css
@@ -122,12 +122,12 @@ input:checked + .slider:before {
 }
 
 .largeLabel {
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
   margin: 0px;
 }
 
 .smallLabel {
-  composes: titleMedium from "../globals.css";
+  composes: titleMedium from "../../styles/globals.css";
   margin: 0px;
 }
 
@@ -136,7 +136,7 @@ input:checked + .slider:before {
   border-radius: 8px;
   background: var(--gray-4);
   color: var(--black);
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
 }
 
 .inputField:disabled {
@@ -196,7 +196,7 @@ input:checked + .slider:before {
 
 .dateInput {
   composes: inputField;
-  composes: labelLarge from "../globals.css";
+  composes: labelLarge from "../../styles/globals.css";
   width: 170px;
   height: 40px;
   margin: 10px 0px;
